+++ tests/cases/cat_pdf/main.sh
@@ -23,11 +23,11 @@ EOD
 mv cat.pdf cat1.pdf
 python3 - << EOD
 from stepup.core.interact import *
-from stepup.reprep.make_manifest import write_manifest
+from stepup.reprep.make_inventory import write_inventory
 watch_delete("cat.pdf")
 run()
 join()
-write_manifest("reproducibility_manifest.txt", ["cat.pdf", "cat1.pdf"])
+write_inventory("reproducibility_inventory.txt", ["cat.pdf", "cat1.pdf"])
 EOD
 
 # Wait for background processes, if any.
@@ -39,4 +39,4 @@ wait
 [[ -f doc2.pdf ]] || exit -1
 [[ -f cat.pdf ]] || exit -1
 [[ -f cat1.pdf ]] || exit -1
-[[ -f reproducibility_manifest.txt ]] || exit -1
+[[ -f reproducibility_inventory.txt ]] || exit -1