+++ tests/cases/make_manifest_in_sub/plan.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-
-from stepup.core.api import static
-from stepup.reprep.api import make_manifest
-
-static("sub/", "sub/MANIFEST.in", "sub/hello.txt")
-make_manifest("sub/MANIFEST.in")